Changes to the documentation.
Largest change was made to bpnr and bpnme. I added a list with all output values for both objects. Please note I did not yet format these using ctrl+shift+/, as this would make list of items and their description very hard to read. This list still needs some definitions to be added and others likely require revision as I may incorrectly interpreted their meaning.

Other changes are minor fixes to inconsistent use of \code{} and formatting. Line 1147 in the file classfunctions.R still requires this but this was the cause of my previously reported error.